Modularize components in LoginLayout
