+++ src/pages/login/index.js
@@ -1,5 +1,5 @@
-import LoginLayout from "components/layouts/LoginLayout";
+import LoginLayout from "components/layouts/LoginLayout/LoginLayout";
 
-export default function Post() {
+export default function Login() {
   return <LoginLayout />;
 }
